+++ JsonToEntity/Business/JavaParseEntity.cs
@@ -1,0 +1,31 @@
+﻿using JsonToEntity.Repository;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace JsonToEntity.Business
+{
+    public class JavaParseEntity : IJavaEntityRepository
+    {
+        public string GetLanguage()
+        {
+            throw new NotImplementedException();
+        }
+
+        public string GetPath()
+        {
+            throw new NotImplementedException();
+        }
+
+        public string GetTemplate()
+        {
+            throw new NotImplementedException();
+        }
+
+        public Dictionary<Type, string> GetTypeMap()
+        {
+            throw new NotImplementedException();
+        }
+    }
+}
