+++ JsonToEntity/Controllers/ValuesController.cs
@@ -7,7 +7,9 @@ using System.Reflection;
 using System.Runtime.Loader;
 using System.Text;
 using System.Threading.Tasks;
+using JsonToEntity.Business;
 using JsonToEntity.Model;
+using JsonToEntity.Repository;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.CSharp;
@@ -22,205 +24,20 @@ namespace JsonToEntity.Controllers
     [ApiController]
     public class ValuesController : ControllerBase
     {
-        private static RazorLightEngine _engine;
-        private static Dictionary<Type, string> _csharpTypeMap;
-        static ValuesController()
-        {
-            _engine = new RazorLightEngineBuilder()
-              .UseFilesystemProject(Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "templates"))
-              .UseMemoryCachingProvider()
-              .Build();
-
-            _csharpTypeMap = new Dictionary<Type, string>
-            {
-                [typeof(byte)] = "byte",
-                [typeof(sbyte)] = "sbyte",
-                [typeof(short)] = "short",
-                [typeof(ushort)] = "unshort",
-                [typeof(int)] = "int",
-                [typeof(uint)] = "uint",
-                [typeof(long)] = "long",
-                [typeof(ulong)] = "ulong",
-                [typeof(float)] = "float",
-                [typeof(double)] = "double",
-                [typeof(decimal)] = "decimal",
-                [typeof(bool)] = "bool",
-                [typeof(string)] = "string",
-                [typeof(char)] = "char",
-                [typeof(Guid)] = "Guid",
-                [typeof(DateTime)] = "DateTime",
-                [typeof(DateTimeOffset)] = "DateTimeOffset",
-                [typeof(TimeSpan)] = "TimeSpan",
-                [typeof(object)] = "object"
-            };
+        private static ICSharpEntityRepository _csharpBusiness;
+        public ValuesController(ICSharpEntityRepository CSharpBusiness) {
+            _csharpBusiness = CSharpBusiness;
         }
-
-        // GET api/values
-        [HttpGet]
-        public ActionResult Get()
-        {
-            // 加载文件
-            var path = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "test", "testclass.cs");
-            var content = System.IO.File.ReadAllText(path);
-            var tree = CSharpSyntaxTree.ParseText(content);
-            var root = (CompilationUnitSyntax)tree.GetRoot();
-
-            // 编译dll
-            var refPaths = new[] {
-                typeof(object).GetTypeInfo().Assembly.Location
-            };
-            MetadataReference[] references = refPaths.Select(r => MetadataReference.CreateFromFile(r)).ToArray();
-            var compilation = CSharpCompilation.Create(
-                Path.GetRandomFileName(),
-                syntaxTrees: new[] { tree },
-                references: references,
-                options: new CSharpCompilationOptions(OutputKind.DynamicallyLinkedLibrary));
-            var semModel = compilation.GetSemanticModel(tree);
-
-            // 解析信息（只包含基本名称、注释信息）
-            var list = new List<ClassInfo>();
-            var ns = root.DescendantNodes().OfType<NamespaceDeclarationSyntax>().First();
-            var ns_symbol = semModel.GetDeclaredSymbol(ns) as INamespaceSymbol;
-            foreach (var item in GetAllTypes(compilation.GetCompilationNamespace(ns_symbol)))
-            {
-                var cinfo = new ClassInfo();
-                cinfo.Name = item.Name;
-                cinfo.RawComment = item.GetDocumentationCommentXml();
-                cinfo.Fields = new List<FieldInfo>();
-                foreach (var member in item.GetMembers().Where(p => p.Kind == SymbolKind.Property))
-                {
-                    var field = new FieldInfo();
-                    field.Name = member.Name;
-                    field.RawComment = member.GetDocumentationCommentXml();
-                    cinfo.Fields.Add(field);
-                }
-                list.Add(cinfo);
-            }
-
-            // 再次解析，回填类型信息
-            using (var ms = new MemoryStream())
-            {
-                EmitResult result = compilation.Emit(ms);
-                if (result.Success)
-                {
-                    ms.Seek(0, SeekOrigin.Begin);
-                    Assembly assembly = AssemblyLoadContext.Default.LoadFromStream(ms);
-                    foreach (var type in assembly.ExportedTypes)
-                    {
-                        var cinfo = list.Find(p => p.Name == type.Name);
-                        foreach (var prop in type.GetProperties())
-                        {
-                            var finfo = cinfo.Fields.Find(p => p.Name == prop.Name);
-                            finfo.RawType = prop.PropertyType;
-                        }
-                    }
-                }
-                else
-                {
-                    IEnumerable<Diagnostic> failures = result.Diagnostics.Where(diagnostic =>
-                        diagnostic.IsWarningAsError ||
-                        diagnostic.Severity == DiagnosticSeverity.Error);
-                    var err = new StringBuilder();
-                    foreach (Diagnostic diagnostic in failures)
-                        err.AppendLine($"\t{diagnostic.Id}: {diagnostic.GetMessage()}");
-                    throw new Exception(err.ToString());
-                }
-            }
-
-            return RenderClass(list);
-        }
-
-        private IEnumerable<INamedTypeSymbol> GetAllTypes(INamespaceSymbol @namespace)
-        {
-            foreach (var type in @namespace.GetTypeMembers())
-                foreach (var nestedType in GetNestedTypes(type))
-                    yield return nestedType;
-
-            foreach (var nestedNamespace in @namespace.GetNamespaceMembers())
-                foreach (var type in GetAllTypes(nestedNamespace))
-                    yield return type;
-        }
-
-        private IEnumerable<INamedTypeSymbol> GetNestedTypes(INamedTypeSymbol type)
-        {
-            yield return type;
-            foreach (var nestedType in type.GetTypeMembers()
-                .SelectMany(nestedType => GetNestedTypes(nestedType)))
-                yield return nestedType;
-        }
-
-        private ContentResult RenderClass(List<ClassInfo> info)
-        {
-            PreProcess(info);
-            var html = InnerRender(info);
+        
+        [HttpGet("csharp")]
+        // POST api/values/csharp
+        public ActionResult CSharpParse() {
             return new ContentResult
             {
                 ContentType = "text/html",
                 StatusCode = (int)HttpStatusCode.OK,
-                Content = html
+                Content = new ParseEntity().CSharpParseEntity(_csharpBusiness)
             };
-        }
-
-        private void PreProcess(List<ClassInfo> list)
-        {
-            MapType(list);
-            FormatComment(list);
-        }
-
-        private void MapType(List<ClassInfo> list)
-        {
-            foreach (var cinfo in list)
-            {
-                foreach (var finfo in cinfo.Fields)
-                {
-                    if (_csharpTypeMap.ContainsKey(finfo.RawType))
-                    {
-                        finfo.TypeStr = _csharpTypeMap[finfo.RawType];
-                    }
-                    else
-                    {
-                        finfo.TypeStr = finfo.RawType.Name;
-                    }
-                }
-            }
-        }
-
-        private void FormatComment(List<ClassInfo> list)
-        {
-            foreach (var cinfo in list)
-            {
-                cinfo.CommentStr = _format(cinfo.RawComment);
-                foreach (var finfo in cinfo.Fields)
-                    finfo.CommentStr = _format(finfo.RawComment);
-            }
-
-            List<string> _format(string raw)
-            {
-                var ret = new List<string>();
-                var arrs = raw.Split(new string[] { "\r\n" }, StringSplitOptions.RemoveEmptyEntries);
-                for (var i = 1; i < arrs.Length - 1; i++)
-                    ret.Add(arrs[i]);
-
-                return ret;
-            }
-        }
-
-        private string InnerRender(List<ClassInfo> model)
-        {
-            var result = string.Empty;
-            var cacheResult = _engine.TemplateCache.RetrieveTemplate("csharp/template.cshtml");
-            if (cacheResult.Success)
-                result = _engine.RenderTemplateAsync(cacheResult.Template.TemplatePageFactory(), model).Result;
-            else
-                result = _engine.CompileRenderAsync("csharp/template.cshtml", model).Result;
-
-            return result;
-        }
-
-        // POST api/values
-        [HttpPost]
-        public void Post([FromBody] string value)
-        {
         }
     }
 }
