+++ JsonToEntity/Repository/ILanguageEntityRepository.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
+
+namespace JsonToEntity.Repository
+{
+    public interface ILanguageEntityRepository
+    {
+        string CSharpParseEntity(ICSharpEntityRepository csharp);
+        string JavaParseEneity(IJavaEntityRepository java);
+    }
+}
