+++ JsonToEntity/Startup.cs
@@ -2,6 +2,8 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using JsonToEntity.Business;
+using JsonToEntity.Repository;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc;
@@ -25,6 +27,7 @@ namespace JsonToEntity
         public void ConfigureServices(IServiceCollection services)
         {
             services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
+            services.AddSingleton<ICSharpEntityRepository, CSharpParseEntity>();
         }
 
         // This method gets called by the runtime. Use this method to configure the HTTP request pipeline.
